Fix remaining Windows + Linux CI failures from dev-rc0 merge
Five build breaks beyond the first cleanup pass:

- CMakeLists.txt: archive OpenSubdiv object files into
  tinyusdz_static via TINYUSDZ_EXT_SOURCES. Previously the
  `$<TARGET_OBJECTS:osd_cpu>` line was commented out, so the
  static lib referenced OpenSubdiv symbols without bundling
  them. With Windows now exercising the OSD path
  (vcsetup*.bat passes -DTINYUSDZ_WITH_OPENSUBDIV=On), this
  surfaced as LNK2019 on tydra_to_renderscene linking.
- scripts/bootstrap-cmake-linux-with-osd.sh: add
  -DTINYUSDZ_WITH_TINYSUBDIV=Off. dev-rc0 made TinySubdiv the
  default ON and the mutual-exclusion check fired during
  CMake configure on the build-osd job.
- src/ascii-parser-basetype.cc: cast `_sr->size()` to size_t
  at the six FindArrayLiteralEnd call sites. uint64_t →
  size_t (uint32_t on 32-bit) tripped
  -Werror,-Wshorten-64-to-32 on the build-32bit job.
- tests/unit/unit-value-view.cc: make the ValueView size
  static_assert pointer-size-aware (16 bytes on 64-bit, 12 on
  32-bit) — matches the header's documented layout. Win32
  MSVC was the only platform where this fired.
- .github/workflows/linux_ci.yml: install libclang-rt-XX-dev
  in the build-arm64-asan and build-32bit prep steps. Both
  bootstraps enable SANITIZE_ADDRESS=1, and the current
  Ubuntu runner images don't ship clang's compiler-rt by
  default — links failed with undefined references to
  __asan_report_load8 / __asan_memmove etc.

+++ src/ascii-parser-basetype.cc
@@ -4043,7 +4043,7 @@ bool AsciiParser::ParseUIntArrayOptimized(std::vector<uint32_t> *result) {
 
   uint64_t start_loc = CurrLoc();
   uint64_t end_loc = start_loc;
-  if (!FindArrayLiteralEnd(_sr->data(), _sr->size(), start_loc, &end_loc)) {
+  if (!FindArrayLiteralEnd(_sr->data(), static_cast<size_t>(_sr->size()), start_loc, &end_loc)) {
     PushError("Unexpected end of input while parsing uint array");
     return false;
   }
@@ -4066,7 +4066,7 @@ bool AsciiParser::ParseInt64ArrayOptimized(std::vector<int64_t> *result) {
 
   uint64_t start_loc = CurrLoc();
   uint64_t end_loc = start_loc;
-  if (!FindArrayLiteralEnd(_sr->data(), _sr->size(), start_loc, &end_loc)) {
+  if (!FindArrayLiteralEnd(_sr->data(), static_cast<size_t>(_sr->size()), start_loc, &end_loc)) {
     PushError("Unexpected end of input while parsing int64 array");
     return false;
   }
@@ -4089,7 +4089,7 @@ bool AsciiParser::ParseUInt64ArrayOptimized(std::vector<uint64_t> *result) {
 
   uint64_t start_loc = CurrLoc();
   uint64_t end_loc = start_loc;
-  if (!FindArrayLiteralEnd(_sr->data(), _sr->size(), start_loc, &end_loc)) {
+  if (!FindArrayLiteralEnd(_sr->data(), static_cast<size_t>(_sr->size()), start_loc, &end_loc)) {
     PushError("Unexpected end of input while parsing uint64 array");
     return false;
   }
@@ -4112,7 +4112,7 @@ bool AsciiParser::ParseTokenArrayOptimized(std::vector<value::token> *result) {
 
   uint64_t start_loc = CurrLoc();
   uint64_t end_loc = start_loc;
-  if (!FindArrayLiteralEnd(_sr->data(), _sr->size(), start_loc, &end_loc)) {
+  if (!FindArrayLiteralEnd(_sr->data(), static_cast<size_t>(_sr->size()), start_loc, &end_loc)) {
     PushError("Unexpected end of input while parsing token array");
     return false;
   }
@@ -4136,7 +4136,7 @@ bool AsciiParser::ParseStringDataArrayOptimized(
 
   uint64_t start_loc = CurrLoc();
   uint64_t end_loc = start_loc;
-  if (!FindArrayLiteralEnd(_sr->data(), _sr->size(), start_loc, &end_loc)) {
+  if (!FindArrayLiteralEnd(_sr->data(), static_cast<size_t>(_sr->size()), start_loc, &end_loc)) {
     PushError("Unexpected end of input while parsing string array");
     return false;
   }
@@ -4160,7 +4160,7 @@ bool AsciiParser::ParseStdStringArrayOptimized(
 
   uint64_t start_loc = CurrLoc();
   uint64_t end_loc = start_loc;
-  if (!FindArrayLiteralEnd(_sr->data(), _sr->size(), start_loc, &end_loc)) {
+  if (!FindArrayLiteralEnd(_sr->data(), static_cast<size_t>(_sr->size()), start_loc, &end_loc)) {
     PushError("Unexpected end of input while parsing std::string array");
     return false;
   }
